+++ packages/@overeng/notion-cli/tsconfig.json.genie.ts
@@ -18,7 +18,5 @@ export default tsconfigJson({
     { path: '../notion-datasource-sync' },
     { path: '../notion-md' },
     { path: '../notion-effect-schema' },
-    { path: '../otel-contract' },
-    { path: '../utils' },
   ],
 } satisfies TSConfigArgs)